Organize zng-ext-image code to be more readable
